Update send_message.js
I was randomly using this package , it worked just fine on catalina, but i got a new machine and tried too use my script again on sonoma, It presented a bug where it worked but it input the number and message into the number input. This was my fix for it! Hope it helps someone out there...